// apps/web-panel-vue/src/views/site/types.d.ts
export interface TrustAsiaTableProps { 
  authDomain: string
  authKey: string
  authPath: string
  authValue: string
  commonName: string
  createTime: number
  dvAuthMethod: string
  endtime: number
  partnerOrderId: string
  product: string | null
  productShortName: string
  setup: boolean
  ssl_id: number
  stateCode: string
  stateName: string
  updateTime: number
  validityPeriod: number
}